+++ app/[lang]/(app)/(home)/(tablet)/restaurant/salons/components/SalonTable.tsx
@@ -28,7 +28,7 @@ import { GrStatusUnknown } from 'react-icons/gr';
 import { useRouter } from 'next/navigation';
 import { useQuery } from '@tanstack/react-query';
 import { getHallKey, getTableOrders } from '../services/salonsApiActions';
-import TableOrders from './TableOrders';
+import TableOrders from './table-orders/TableOrders';
 import { useOrderRedirectLink } from '../hooks/useOrderRedirectLink';
 import { TableUtils } from '../utils/tableUtils';
 import { IoAlbums } from 'react-icons/io5';
@@ -63,7 +63,6 @@ export default function SalonTable({
  const { data: ordersList, isLoading: isLoadingOrdersList } = useQuery({
   enabled: showTableOrdersList && tableUtils.tableType === 'normal',
   queryKey: [getHallKey, 'ordersList', table.tableID.toString()],
-  gcTime:0,
   async queryFn({ signal }) {
    const res = await getTableOrders({
     tableID: table.tableID,
@@ -368,7 +367,7 @@ export default function SalonTable({
       </DrawerHeader>
       <div className='overflow-hidden overflow-y-auto p-4'>
        <TableOrders
-        dic={tableUtils.dic}
+        dic={tableUtils.dic['multiOrder']}
         data={ordersList}
         isLoading={isLoadingOrdersList}
         orderRedirectLink={orderRedirectLink}